Synthetic version of MDS long-term data

Synthetic version of MDS long-term data

dat_mds_synth

Format

ci_s_allo1

Competing event indicator: 0 = censored, 1 = relapse, 2 = non-relapse mortality.

ci_allo1

Time to first competing event (years).

match_allo1_1

Sex match patient and donor. Unordered categorical with four levels, corresponding to all combinations of recipient male/female and donor male/female.

mdsclass

MDS groups based on subclassification at alloHCT. Categorical with three levels: MDS without excess blasts, with excess blast and sAML.

donorrel

Human leukocyte antigen match between patient and donor. Binary: Identical sibling or Other.

karnofsk_allo1

Karnofsky performance status. Three-level ordered categorical: >=90, 80 and <=70.

crnocr

Stage at alloHCT. Three-level unordered categorical: CR, noCR and Untreated/not aimed at remission.

cmv_combi_allo1_1

Cytomegalovirus (CMV) status in patient and donor. Four-level categorical corresponding to all combinations of patient positive/negative and donor positive/negative. For example, level "+/-" corresponds to CMV positive in patient and negative in donor.

cytog_threecat

Cytogenetics categories for International Prognostic Scoring System (IPSS-R) - see Greenberg et al. (2012).

hctci_risk

Hematopoietic stemcell transplantation-comorbidity index score. Three-level ordered categorical: low risk (0), intermediate risk (1-2) and high risk (>= 3). See Sorror et al. (2005).

agedonor_allo1_decades

Donor age at alloHCT (decades).

age_allo1_decades

Patient age at alloHCT (decades).

References

Greenberg, Peter L., Heinz Tuechler, Julie Schanz, Guillermo Sanz, Guillermo Garcia-Manero, Francesc Solé, John M. Bennett et al. "Revised international prognostic scoring system for myelodysplastic syndromes." Blood 120, no. 12 (2012): 2454-2465.Sorror, Mohamed L., Michael B. Maris, Rainer Storb, Frederic Baron, Brenda M. Sandmaier, David G. Maloney, and Barry Storer. "Hematopoietic cell transplantation (HCT)-specific comorbidity index: a new tool for risk assessment before allogeneic HCT." Blood 106, no. 8 (2005): 2912-2919.