Generate competing risks times + event indictor (2 events)

Generate competing risks times + event indictor (2 events)

gen_cmprsk_times(n, dat, ev1_pars, ev2_pars, rate_cens, mod_type = "latent")

Arguments

n
Sample size.

dat
Data frame containing n rows with X and Z as columns

ev1_pars
Named list of ("a1", "h1_0", "b1", "gamm1")

ev2_pars
Named list of ("a2", "h2_0", "b2", "gamm2")

rate_cens
Rate of exponential distributiion for censoring. If 0 means no censoring is applied.

mod_type
Either "latent", weibull times generated from separate weibull distribution, or "total", times generated from sum of cause-specific hazards (using inverse transform method). For educational purposes - both methods yield virtually same results.

Value

Data-frame with missings induced