Custom ggplot-based nested loop plots

Custom ggplot-based nested loop plots

ggplot_nlp(
  dat,
  estim,
  method_var,
  true,
  step_factors,
  pointsize = 2,
  point_dodge = 0.75,
  text_size = 4,
  top_step = NULL,
  height_betw_steps = NULL,
  height_steps = NULL,
  step_labels = NULL
)

Arguments

dat
Data frame with simulation results

estim
Character for column containing estimates

method_var
Character for column containing method

true
Character for column containing true data-generating estimand value, or a scalar e.g. 0

step_factors
Vector of variable names to use as step factors at bottom of the NLP.

pointsize
Size of points

point_dodge
Measure of horizontal dodge

text_size
Size of the step labels

top_step
Position of top step (on scale of data)

height_betw_steps
Height between NLP step (on scale of data)

height_steps
Height of the NLP steps themselves (on scale of data)

step_labels
Optional named vector of labels for the step functions. Defaults to variable name and the factor labels

Value

A ggplot2 object

Examples

sim_results <- CauseSpecCovarMI::regr_results

# Plot example code here...