Summarised simulation results - predictions

Summarised simulation results - predictions

preds_results

Format

analy

Method indicator, please refer to manuscript for abbreviations used

m

Number of imputed datasets

combo-X_Z

Covariate combination defining reference patient for which we predict

times

Prediction horizon (in years)

state

State for which we estimate probability at times

true

True data-generating cumulative incidence at a horizon, and for a certain reference patient

scen_num

Number label for scenario

n

Sample size in simulated datasets

prob

Estimated probability

emp_se

Empirical SE of the probabilities

bias

Bias

rmse

Coverage

rmse_mcse

Coverage

bias_mcse

Monte carlo error of bias

prop_miss

Proportion of missing values in X

beta1

Data-generating ffect of X on event 1

miss_mech

Missingness mechanism

X_level

Measurement level of X

rho

Correlation between X and Z

eta1

Strength of missingness mechanism for non-MCAR scenarios

haz_shape

Shapes of baseline hazard for competing events, either "similar" or "different"